moved plotting function to new file
